split backendmap to own file
